+++ CHANGES.md
@@ -9,6 +9,7 @@ The version numbers are in the form `MAJOR.MINOR.PATCH`, where major releases in
 
 * **Changed**
   * Update default solver settings for incompressible elasticity for Schur factoriztion from `full` to `lower`.
+  * Set the default length scale to 1 km instead of 100 km for consistency with most quasi-static problem with faults. For simulations with faults, the length scale should be set to the discretization size on the fault; this is related to the preconditioner which links the length scale to the discretization size.
 * **Added**
   * `examples/subduction-3d`: Add expected Gmsh output for geometric operations.
 * **Fixed**